Move history management to execution_context (WIP, have not been run/tested).
Execute context will manage list of "markers" in browser history, and will try to automatically "go back" when element from an older page in the history is required.

I abandoned this approach because:
1) The approach is complex and potential over-engineered;
2) There might be performance issues - too many scripts to run on the browser;
3) I'm not sure it will work in all cases.

The alternative is to be explicit, so the script writer will need to tell the browser to "to back" when needed. We can also think of some heuristics that can do that automatically in some cases. It seems, at this point, that the alternative is simpler.